requirements.txt
langchain==0.1.12
langchain_community==0.0.28
langchain_openai==0.0.8
python-dotenv==1.0.1
streamlit==1.32.2
